renderization components reorganized
